requirements.txt
flask
flask-login
pymongo
werkzeug
python-dotenv
